+++ viewer/src/store.js
@@ -4,6 +4,7 @@ import createSagaMiddleware from 'redux-saga';
 
 import rootReducer from './reducers/rootReducer';
 import rootSaga from './sagas/rootSaga';
+import initHotkeys from './hotkeys';
 
 const sagaMiddleware = createSagaMiddleware();
 const configureStore = () => {
@@ -12,6 +13,7 @@ const configureStore = () => {
         sagaMiddleware
     ));
     sagaMiddleware.run(rootSaga);
+    initHotkeys(store);
     return store;
 };